requirements.txt
Django
django
deepgram-sdk
python-decouple
aiohttp
python-dotenv
deepgram-sdk
channels==3.0.4
openai
pinecone-client
pandas
typing
tqdm
langchain
wget
ajax